+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.0] - 2022.07.27
 
-(may not have any breaking changes)
 - update all dependencies to latest
 - fix rml:reference with rr:termType rr:IRI
+- remove in-mapping JS functions & http calls (`jsFunction`, `httpCall`), use normal functions instead
+
 
 ## [1.14.2] - 2022.07.27